properly calculated quorum

+++ internal/vote/vote.go
@@ -12,7 +12,7 @@ type Proposal struct {
 	VotedNo     int
 	Veto        int
 	DeadlineHrs float64
-	Voted       int
+	Voted       float64
 }
 
 //go:generate mockgen -source vote.go -destination vote_mock.go -package vote
